projects/proj-1/requirements.txt
assertpy~=1.1
parameterized
setuptools~=40.8.0
pyhamcrest
nose